# public/swagger/web/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Patient Tracker API",
        "version": "1.0"
    },
    "paths": {
...
        "/patients/findAll": {
            "get": {
                "tags": [
                    "Patients"
                ],
                "summary": "Find all patients",
                "operationId": "patients_findAll",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommonFindAll"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No result found"
                    }
                }
            }
        },
...
        "/patients/update/{patientId}": {
            "put": {
                "tags": [
                    "Patients"
                ],
                "summary": "Patient",
                "operationId": "patients_update",
                "parameters": [
                    {
                        "name": "patientId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update user",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PatientSave"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request, returns to validation errors"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CommonFindAll": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "string"
                                },
                                "name": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "ObjectId": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "PatientSave": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "userId": {
                        "$ref": "#/components/schemas/ObjectId"
                    },
                    "gender": {
                        "$ref": "#/components/schemas/ObjectId"
                    },
                    "ageGroup": {
                        "$ref": "#/components/schemas/ObjectId"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}